Update - Move to Poetry for packaging - Update some code
