+++ inferius/devices.json
