+++ inferius/manifest.py
