+++ pyproject.toml
@@ -1,0 +1,45 @@
+[tool.poetry]
+name = "inferius"
+version = "0.1.0"
+description = "A Python CLI tool for creating & restoring custom firmwares to an iOS device."
+authors = ["m1stadev <adamhamdi31@gmail.com>"]
+license = "MIT"
+readme = "README.md"
+repository = "https://github.com/m1stadev/Inferius"
+keywords = ["restore", "pwndfu", "ios", "checkm8"]
+classifiers = [
+    "Development Status :: 4 - Beta",
+    "Environment :: Console",
+    "Programming Language :: Python :: 3.7",
+    "Topic :: Utilities"
+]
+packages = [
+    { include = "inferius" }
+]
+
+[tool.poetry.scripts]
+pyimg4 = "inferius.__main__:main"
+
+[tool.poetry.dependencies]
+python = "^3.7"
+pyusb = "^1.2.1"
+remotezip = "^0.9.4"
+requests = "^2.27.1"
+
+[tool.poetry.dev-dependencies]
+black = {version = "^22.1.0", python = "^3.7"}
+isort = {version = "^5.10.1", python = "^3.7"}
+
+[tool.black]
+skip-string-normalization = true
+
+[tool.isort]
+profile = "black"
+src_paths = ["inferius"]
+
+[tool.poetry.urls]
+"Bug Tracker" = "https://github.com/m1stadev/Inferius/issues"
+
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
