+++ requirements.txt
@@ -2,4 +2,5 @@ bsdiff4
 mwclient
 remotezip
 requests
+pymobiledevice3
 pyusb
